2IS063/requirements2.txt
pandas
textmining3
wordcloud
matplotlib
scikit-learn
numpy
nltk
gensim
pyLDAvis
tqdm
langdetect
seaborn
